Move dashboard down into environment
This allows it to share IO actions with chart. It makes more sense
anyway to have the environment manage its own states.

Also, moving everything under a single containing div seems to have
flipped some magic layerization flag in chrome, because now the recipes
modal is buttery smooth. Wins all around!
